+Update +Delete +Responses

+++ backend/routes/users.py
@@ -14,7 +14,7 @@ import logging
 
 router = APIRouter(prefix="/users")
 
-@router.post("/create", response_model=UsersPublic, name="users:create-user", 
+@router.post("/", response_model=UsersPublic, name="users:create-user", 
              responses={
                  status.HTTP_400_BAD_REQUEST: {"model": UsersPublic, "description": "User already exists", },
                  status.HTTP_201_CREATED: {"model": UsersPublic, "description": "User created"},
@@ -35,39 +35,53 @@ async def create_user(
     return JSONResponse(status_code=status.HTTP_201_CREATED, content=created_user.model_dump())
 
 
-@router.get("/{user_id}", response_model=UsersPublic, name="users:get-user", status_code=status.HTTP_302_FOUND)
+@router.get("/{user_id}", response_model=UsersPublic, name="users:get-user",
+            responses={
+                status.HTTP_404_NOT_FOUND: {"description": "User not found"},
+                status.HTTP_200_OK: {"model": UsersPublic, "description": "User found"}
+                })
 async def get_user_by_id(
     user_id: int,
     users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
 ):
     user = await users_repo.get_user_by_id(user_id=user_id)
     if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
-    return user
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"detail": "User not found"})
+    return JSONResponse(status_code=status.HTTP_200_OK, content=user.model_dump())
 
 
-# @router.patch("/{user_id}", response_model=UsersPublic)
-# async def update_user(
-#     user_id: int,
-#     user_update: UsersUpdate,
-#     users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
-# ):
-#     existing_user = await users_repo.get_user_by_id(user_id=user_id)
-#     if not existing_user:
-#         raise HTTPException(status_code=404, detail="User not found")
+@router.patch("/{user_id}", response_model=UsersPublic, name="users:update-user",
+              responses={
+                  status.HTTP_404_NOT_FOUND: {"description": "User not found"},
+                  status.HTTP_200_OK: {"model": UsersPublic, "description": "User updated"}
+                  })
+async def update_user(
+    user_id: int,
+    user_update: UsersUpdate,
+    users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
+):
+    user = await users_repo.get_user_by_id(user_id=user_id)
+    if not user:
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"detail": "User not found"})
     
-#     updated_user = await users_repo.update_user(user_id=user_id, user_update=user_update)
-#     return updated_user
+    updated_user = await users_repo.update_user(user_id=user_id, user_update=user_update)
+    return JSONResponse(status_code=status.HTTP_200_OK, content=updated_user.model_dump())
 
 
-@router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
+@router.delete("/{user_id}", 
+               responses={
+                   status.HTTP_404_NOT_FOUND: {"description": "User not found"},
+                   status.HTTP_200_OK: {"description": "User deleted"}
+               })
 async def delete_user(
     user_id: int,
     users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
 ):
-    deleted = await users_repo.delete_user(user_id=user_id)
-    if not deleted:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    user = await users_repo.get_user_by_id(user_id=user_id)
+    if not user:
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"detail": "User not found"})
+    await users_repo.delete_user(user_id=user_id)
+    return JSONResponse(status_code=status.HTTP_200_OK, content={"detail": "User deleted"})
 
 # @router.post("/{user_id}/documents/upload", response_model=UsersDocumentUploadResponse, status_code=status.HTTP_200_OK)
 # async def upload_document(
